chore: remove unnecessary code
