+++ backend/lib/request_handlers/unimplemented_handler.dart
@@ -9,7 +9,7 @@ import 'package:dart_frog/dart_frog.dart';
 
 Future<Response> unimplementedHandler([RequestContext? context]) async {
   return Response.json(
-    body: {'error': '👀 Looks like you are lost 🔦'},
+    body: {'error': '👀 Not implemented yet'},
     statusCode: HttpStatus.notImplemented,
   );
 }
